Component seperation into seperate files
